requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
psycopg2-binary
pandas
numpy
PyYAML
pydantic
neo4j

